pyproject.toml
[project]
name = "autotrader-pro"
version = "1.0.0"
description = "Professional Automated Trading System with ML and Advanced Risk Management"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core dependencies
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    # Async and scheduling
    "apscheduler>=3.10.0",
    "aiofiles>=23.0.0",
    # Database
    "sqlmodel>=0.0.14",
    "alembic>=1.12.0",
    # Data analysis and ML
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "ta>=0.10.0",
    # Financial data
    "yfinance>=0.2.0",
    "alpaca-py>=0.7.0",
    # HTTP and API
    "httpx>=0.25.0",
    "requests>=2.31.0",
    "requests-cache>=1.1.0",
    "requests-ratelimiter>=0.4.0",
    # Web framework
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    # Configuration and utilities
    "pyyaml>=6.0",
    "arrow>=1.2.3",
    "pendulum>=3.1.0",
    # Logging
    "structlog>=23.0.0",
    "json-log-formatter>=0.5.0",
    # Dashboard and visualization
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "dash>=2.14.0",
    "dash-bootstrap-components>=1.5.0",
    "prophet>=1.1.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0",
]

dashboard = [
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
]

broker = [
    "alpaca-trade-api>=3.0.0",
    "requests>=2.31.0",
]

[project.scripts]
autotrader = "src.main_automated:main"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "--strict-markers --disable-warnings"

[tool.uv]
dev-dependencies = [
    "hypothesis",
    "pre-commit>=3.0.0",
    "pytest>=7.0.0",
    "pytest-mock",
    "ruff",
    "black>=23.0.0",
    "mypy>=1.0.0",
]
